move the form of creating car from CreateCar.js to component/CreateCarForm.js
